Sending Data Over Wifi
Did a bit of organizing of files
Able to send IMU data over UDP to my computer and run the python script to determine what kind of punch it is. Need to work a bit more on the analysis and how to incorporate it together with the rest of the code.